Another bug fixed: Incorrect method
I have changed the .innerHTML in the code to textContent. 

With innerHTML there is a security issue because innerHTML processes and interprets HTML tags, it is advisable to use it only when inserting content from trusted sources
textContent property ignores all HTML tags and returns only the text. textContent reads text as it is in the markup and is a more efficient choice when it comes to performance